call autosave service immediatelly in order for it to be available for cancelling immediatelly as well.
