+++ scripts/apps/authoring/authoring/services/AutosaveService.ts
@@ -1,7 +1,11 @@
 import * as helpers from 'apps/authoring/authoring/helpers';
+import {applyMiddleware as coreApplyMiddleware} from 'core/middleware';
+import {onChangeMiddleware} from '../index';
+import {extensions} from 'appConfig';
+import {IArticle} from 'superdesk-api';
 
-const RESOURCE = 'archive_autosave',
-    AUTOSAVE_TIMEOUT = 3000;
+const RESOURCE = 'archive_autosave';
+const AUTOSAVE_TIMEOUT = 3000;
 
 let $q, $timeout, api;
 
@@ -41,7 +45,7 @@ export class AutosaveService {
     /**
      * Auto-saves an item
      */
-    save(item, orig, timeout = AUTOSAVE_TIMEOUT) {
+    save(item: IArticle, orig: IArticle, timeout: number = AUTOSAVE_TIMEOUT) {
         if (!item._editable || !item._locked) {
             return $q.reject('item not ' + item._editable ? 'locked' : 'editable');
         }
@@ -51,13 +55,34 @@ export class AutosaveService {
         let id = item._id;
 
         this.timeouts[id] = $timeout(() => {
-            var diff = helpers.extendItem({_id: id}, item);
+            coreApplyMiddleware(onChangeMiddleware, {item: item, original: orig}, 'item')
+                .then(() => {
+                    const onUpdateFromExtensions = Object.values(extensions).map(
+                        (extension) => extension.activationResult?.contributions?.authoring?.onUpdate,
+                    ).filter((updateFn) => updateFn != null);
 
-            helpers.filterDefaultValues(diff, orig);
-            return api.save(RESOURCE, {}, diff).then((_autosave) => {
-                orig._autosave = _autosave;
-                return _autosave;
-            });
+                    return (
+                        onUpdateFromExtensions.length < 1
+                            ? Promise.resolve(item)
+                            : onUpdateFromExtensions
+                                .reduce(
+                                    (current, next) => current.then(
+                                        (result) => next(orig, result),
+                                    ),
+                                    Promise.resolve(item),
+                                )
+                                .then((nextItem) => angular.extend(item, nextItem))
+                    ).then((itemLatest: IArticle) => {
+                        var diff = helpers.extendItem({_id: id}, itemLatest);
+
+                        helpers.filterDefaultValues(diff, orig);
+
+                        return api.save(RESOURCE, {}, diff).then((_autosave) => {
+                            orig._autosave = _autosave;
+                            return _autosave;
+                        });
+                    });
+                });
         }, timeout, false);
 
         return this.timeouts[id];
